Replace poisson implementation
Solves an issue with the knuth poisson algorithm for large lambdas around 700
and above, where sampled values could only grow to around 700.

Essentially for large lambdas the negative exponent becomes very small and
underflows to zero, which means the loop returns very small values.

This change uses the normal approximation to the Poisson distribution for
lambdas >= 50, chosen because a gaussian variable is an extremely
close approximation of poisson for lambda values >= 50.